requirements.txt
sounddevice==0.5.1
numpy==2.0.2
scipy==1.13.1
pynput==1.7.7
openai>=1.13.0
python-dotenv==1.0.1 